Reload on SIGHUP
When qwfwd receives a SIGHUP signal, it signals the application to
reload the configuration. In effect, SIGHUP triggers "exec qwfwd.cfg".
This functionality is not supported on Windows.

Additionally, SIGHUP purges the entire whitelist so that the
configuration reload can reinitialize the addresses in the whitelist.

It is also possible to purge the whitelist manually by running the
"whitelistpurge" command.